+++ opentelemetry-distro/src/opentelemetry/distro/__init__.py
@@ -14,15 +14,19 @@
 #
 import os
 from logging import getLogger
+from os import environ
 from typing import Sequence, Tuple
 
 from pkg_resources import iter_entry_points
 
 from opentelemetry import trace
-from opentelemetry.configuration import Configuration
+from opentelemetry.environment_variables import (
+    OTEL_PYTHON_IDS_GENERATOR,
+    OTEL_PYTHON_SERVICE_NAME,
+    OTEL_TRACE_EXPORTER,
+)
 from opentelemetry.instrumentation.configurator import BaseConfigurator
 from opentelemetry.instrumentation.distro import BaseDistro
-from opentelemetry.sdk.metrics.export import MetricsExporter
 from opentelemetry.sdk.resources import Resource
 from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import (
@@ -36,34 +40,40 @@ logger = getLogger(__file__)
 
 EXPORTER_OTLP = "otlp"
 EXPORTER_OTLP_SPAN = "otlp_span"
-EXPORTER_OTLP_METRIC = "otlp_metric"
 
 RANDOM_IDS_GENERATOR = "random"
 _DEFAULT_IDS_GENERATOR = RANDOM_IDS_GENERATOR
 
 
 def _get_ids_generator() -> str:
-    return Configuration().IDS_GENERATOR or _DEFAULT_IDS_GENERATOR
+    return environ.get(OTEL_PYTHON_IDS_GENERATOR, _DEFAULT_IDS_GENERATOR)
 
 
 def _get_service_name() -> str:
-    return Configuration().SERVICE_NAME or ""
+    return environ.get(OTEL_PYTHON_SERVICE_NAME, "")
 
 
 def _get_exporter_names() -> Sequence[str]:
-    exporter = Configuration().EXPORTER or EXPORTER_OTLP
-    if exporter.lower().strip() == "none":
-        return []
+    trace_exporters = environ.get(OTEL_TRACE_EXPORTER)
 
-    names = []
-    for exp in exporter.split(","):
-        name = exp.strip()
-        if name == EXPORTER_OTLP:
-            names.append(EXPORTER_OTLP_SPAN)
-            names.append(EXPORTER_OTLP_METRIC)
-        else:
-            names.append(name)
-    return names
+    exporters = set()
+
+    if (
+        trace_exporters is not None
+        or trace_exporters.lower().strip() != "none"
+    ):
+        exporters.update(
+            {
+                trace_exporter.strip()
+                for trace_exporter in trace_exporters.split(",")
+            }
+        )
+
+    if EXPORTER_OTLP in exporters:
+        exporters.pop(EXPORTER_OTLP)
+        exporters.add(EXPORTER_OTLP_SPAN)
+
+    return list(exporters)
 
 
 def _init_tracing(
@@ -113,8 +123,8 @@ def _import_tracer_provider_config_components(
 
 def _import_exporters(
     exporter_names: Sequence[str],
-) -> Tuple[Sequence[SpanExporter], Sequence[MetricsExporter]]:
-    trace_exporters, metric_exporters = {}, {}
+) -> Sequence[SpanExporter]:
+    trace_exporters = {}
 
     for (
         exporter_name,
@@ -124,15 +134,11 @@ def _import_exporters(
     ):
         if issubclass(exporter_impl, SpanExporter):
             trace_exporters[exporter_name] = exporter_impl
-        elif issubclass(exporter_impl, MetricsExporter):
-            metric_exporters[exporter_name] = exporter_impl
         else:
             raise RuntimeError(
-                "{0} is neither a trace exporter nor a metric exporter".format(
-                    exporter_name
-                )
+                "{0} is not a trace exporter".format(exporter_name)
             )
-    return trace_exporters, metric_exporters
+    return trace_exporters
 
 
 def _import_ids_generator(ids_generator_name: str) -> IdsGenerator:
@@ -151,19 +157,10 @@ def _import_ids_generator(ids_generator_name: str) -> IdsGenerator:
 
 def _initialize_components():
     exporter_names = _get_exporter_names()
-    trace_exporters, metric_exporters = _import_exporters(exporter_names)
+    trace_exporters = _import_exporters(exporter_names)
     ids_generator_name = _get_ids_generator()
     ids_generator = _import_ids_generator(ids_generator_name)
     _init_tracing(trace_exporters, ids_generator)
-    # We don't support automatic initialization for metric yet but have added
-    # some boilerplate in order to make sure current implementation does not
-    # lock us out of supporting metrics later without major surgery.
-    _init_metrics(metric_exporters)
-
-
-def _init_metrics(exporters: Sequence[MetricsExporter]):
-    if exporters:
-        logger.warning("automatic metric initialization is not supported yet.")
 
 
 class Configurator(BaseConfigurator):
@@ -178,4 +175,4 @@ class OpenTelemetryDistro(BaseDistro):
     """
 
     def _configure(self, **kwargs):
-        os.environ.setdefault("OTEL_EXPORTER", "otlp")
+        os.environ.setdefault(OTEL_TRACE_EXPORTER, "otlp_span")