+++ opentelemetry-distro/tests/test_distro.py
@@ -19,6 +19,7 @@ from unittest import TestCase
 from pkg_resources import DistributionNotFound, require
 
 from opentelemetry.distro import OpenTelemetryDistro
+from opentelemetry.environment_variables import OTEL_TRACE_EXPORTER
 
 
 class TestDistribution(TestCase):
@@ -30,6 +31,6 @@ class TestDistribution(TestCase):
 
     def test_default_configuration(self):
         distro = OpenTelemetryDistro()
-        self.assertIsNone(os.environ.get("OTEL_EXPORTER"))
+        self.assertIsNone(os.environ.get(OTEL_TRACE_EXPORTER))
         distro.configure()
-        self.assertEqual("otlp", os.environ.get("OTEL_EXPORTER"))
+        self.assertEqual("otlp_span", os.environ.get(OTEL_TRACE_EXPORTER))