+++ opentelemetry-sdk/tests/conftest.py
@@ -14,12 +14,14 @@
 
 from os import environ
 
+from opentelemetry.environment_variables import OTEL_PYTHON_CONTEXT
+
 
 def pytest_sessionstart(session):
     # pylint: disable=unused-argument
-    environ["OTEL_CONTEXT"] = "contextvars_context"
+    environ[OTEL_PYTHON_CONTEXT] = "contextvars_context"
 
 
 def pytest_sessionfinish(session):
     # pylint: disable=unused-argument
-    environ.pop("OTEL_CONTEXT")
+    environ.pop(OTEL_PYTHON_CONTEXT)