+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/open-telemetry/opentelemetry-python/compare/v0.17b0...HEAD)
 
+### Changed
+- Tracer and Meter provider environment variables are now consistent with the rest
+  ([#1571](https://github.com/open-telemetry/opentelemetry-python/pull/1571)])
+
 ### Added
 - Added `end_on_exit` argument to `start_as_current_span`
   ([#1519](https://github.com/open-telemetry/opentelemetry-python/pull/1519)])
@@ -15,6 +19,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#1480](https://github.com/open-telemetry/opentelemetry-python/pull/1480))
 - Allow missing carrier headers to continue without raising AttributeError
   ([#1545](https://github.com/open-telemetry/opentelemetry-python/pull/1545))
+
+
+### Changed
+- Read-only Span attributes have been moved to ReadableSpan class
+  ([#1560](https://github.com/open-telemetry/opentelemetry-python/pull/1560))
+
+### Removed
+- Remove Configuration
+  ([#1523](https://github.com/open-telemetry/opentelemetry-python/pull/1523))
+- Remove Metrics as part of stable, marked as experimental
+  ([#1568](https://github.com/open-telemetry/opentelemetry-python/pull/1568))
 
 ## [0.17b0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v0.17b0) - 2021-01-20
 
@@ -79,6 +94,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#1536](https://github.com/open-telemetry/opentelemetry-python/pull/1536))
 - Fix TraceState to adhere to specs
   ([#1502](https://github.com/open-telemetry/opentelemetry-python/pull/1502))
+- Update Resource `merge` key conflict precedence
+  ([#1544](https://github.com/open-telemetry/opentelemetry-python/pull/1544))
 
 ### Removed
 - `opentelemetry-api` Remove ThreadLocalRuntimeContext since python3.4 is not supported.